fix: prevent jitter truncation and phantom arrivals from empty routes
- Fix departure jitter calculation to use modulo on full usize suffix before narrowing to u16, preventing silent distribution shifts for large sim ordinals (> u16::MAX)
- Add defensive guard in trip tick to handle empty route plans not at destination, marking them as unserved instead of recording phantom arrivals
- Add regression tests for both fixes

+++ crates/caelum-core/src/commute.rs
@@ -42,7 +42,11 @@ pub fn departure_minute_for_sim(sim_id: &str, template: &str, direction: &str) -
         _ => (420, 540),
     };
     let span = end - start;
-    let jitter = numeric_id_suffix(sim_id) as u16 % (span + 1);
+    // Take the modulo on the full `usize` suffix, then narrow. Casting to `u16` first
+    // would truncate suffixes > u16::MAX before the modulo, silently shifting the jitter
+    // distribution for very large sim ordinals. The modulo result is always <= span
+    // (<= 120), so the final `as u16` cannot truncate.
+    let jitter = (numeric_id_suffix(sim_id) % usize::from(span + 1)) as u16;
     start + jitter
 }
